requirements.txt
requests~=2.32.4
google-genai
Django~=5.2.4
python-dotenv~=1.1.1
firebase-admin~=6.2.0
pytz~=2024.1